+++ packages/extractor/src/index.ts
@@ -3,3 +3,4 @@ export * from "./parser.js";
 export * from "./hashing.js";
 export * from "./constants.js";
 export * from "./types.js";
+export * from "./targets.js";